[MAINT] move nilearn.plotting.image.utils
