+++ nilearn/datasets/tests/test_utils.py
@@ -16,6 +16,8 @@ import requests
 
 from nilearn.datasets import _utils
 from nilearn.datasets.tests.conftest import Response
+from nilearn.datasets.utils import MNI152TEMPLATE, load_mni152_template
+from nilearn.image import get_data, reorder_img
 
 datadir = _utils.PACKAGE_DIRECTORY / "data"
 
@@ -617,3 +619,11 @@ def test_naive_ftp_adapter():
         resp = sender.send(
             requests.Request("GET", "ftp://example.com").prepare()
         )
+
+
+def test_mni152template_is_reordered():
+    """See issue #2550."""
+    reordered_mni = reorder_img(load_mni152_template(resolution=2))
+    assert np.allclose(get_data(reordered_mni), get_data(MNI152TEMPLATE))
+    assert np.allclose(reordered_mni.affine, MNI152TEMPLATE.affine)
+    assert np.allclose(reordered_mni.shape, MNI152TEMPLATE.shape)
